Fix shader cache leak (#1580)
* Fix compilation issue in Texture2D.cpp with _AX_DEBUG=2.

* Fix memory leak: ShaderCache was not purged.

The ShaderCache class had two static members: a map for the cache, and
a pointer to the single instance of ShaderCache. The map was
supposedly cleared in the destructor but there was actually no
instance of ShaderCache (getInstance() was never called and the other
functions accessed the map directly). This commit removes the
singleton "interface" such that only the map is used, then it cleared
the map upon request.