Removed Node::removeFromParentAndCleanup
This change removes the Node::removeFromParentAndCleanup function, and instead gives Node::removeFromParent and optional bool for if the node should be cleaned up or not.

This defaults to true, since the original intention was that it would cleanup the memory by default, and only not cleanup the memory when told to do so by the programmer with the old Node::removeFromParentAndCleanup function.

I've also added missing licenses for the Inspector.cpp/.h files, and the other C++ and header files I had to modify that were using the old function.