Missed a couple Node::removeFromParentAndCleanup calls
This also (hopefully, I don't have time to test it) LUA support for this refactor.

I also removed the license info I added to Inspector.cpp/.h since I realized it was in the extensions folder, and so it didn't have any info to begin with. Not sure if that was intentional or not, so I just removed it to play it safe.